+++ geode-for-redis/src/main/java/org/apache/geode/redis/internal/data/NullRedisList.java
@@ -38,7 +38,7 @@ class NullRedisList extends RedisList {
   }
 
   @Override
-  public long linsert(byte[] elementToInsert, byte[] referenceElement, boolean before,
+  public int linsert(byte[] elementToInsert, byte[] referenceElement, boolean before,
       Region<RedisKey, RedisData> region, RedisKey key) {
     return 0;
   }
